+++ BattleCity/Client/mainWindow.py
@@ -231,8 +231,8 @@ class MainWindow(QMainWindow):
     def center(self):
         screen = QDesktopWidget().screenGeometry()
         size = self.geometry()
-        self.move((screen.width() - size.width()) / 2,
-                  (screen.height() - size.height()) / 2 - 100)
+        self.move(int((screen.width() - size.width()) / 2),
+            int((screen.height() - size.height()) / 2 - 100))
 
     def closeEvent(self, *args, **kwargs):
         oalQuit()
